Put trace_test_utils behind a test-utils feature flag
To reduce the possibility of test helper functions being used in
non-test code. Also, move create_send_data and poll_for_mock_hit
functions into trace_test_utils.
