+++ trace-mini-agent/Cargo.toml
@@ -23,3 +23,4 @@ rmp-serde = "1.1.1"
 serial_test = "2.0.0"
 duplicate = "0.4.1"
 tempfile = "3.3.0"
+datadog-trace-utils = { path = "../trace-utils", features=["test-utils"] }
