+++ trace-utils/Cargo.toml
@@ -23,8 +23,12 @@ datadog-trace-normalization = { path = "../trace-normalization" }
 tokio = { version = "1", features = ["macros", "rt-multi-thread"] }
 rand = "0.8.5"
 bytes = "1.6.0"
+# This should only be used for testing. It isn't under dev-dependencies because test-utils can't be under #[cfg(test)].
+httpmock = {  version = "0.7.0", optional = true}
 
 [dev-dependencies]
 tokio = { version = "1", features = ["macros", "rt-multi-thread"] }
-httpmock = "0.7.0"
 serde_json = "1.0"
+
+[features]
+test-utils = ["httpmock"]
